Fix issues from CodeFactor
