+++ Zilon.Core/Zilon.Core/Tactics/StaticObject.cs
@@ -81,22 +81,4 @@ namespace Zilon.Core.Tactics
             }
         }
     }
-
-    public static class IStaticObjectExtensions
-    {
-        public static TStaticObjectModule GetModuleSafe<TStaticObjectModule>(this IStaticObject source)
-        {
-            if (source is null)
-            {
-                throw new ArgumentNullException(nameof(source));
-            }
-
-            if (!source.HasModule<TStaticObjectModule>())
-            {
-                return default;
-            }
-
-            return source.GetModule<TStaticObjectModule>();
-        }
-    }
 }
